separate helpers and proof utils
