+++ lib/utils/proofs.js
@@ -11,100 +11,11 @@
  * limitations under the License.
  */
 
-import url from 'url'
 import cpp from 'chainpoint-parse'
-import uuidValidate from 'uuid-validate'
 import uuidv1 from 'uuid/v1'
-import { isURL, isIP, isJSON, isBase64 } from 'validator'
+import { isJSON, isBase64 } from 'validator'
 import { isEmpty, isString, isArray, has, isObject, forEach, map, isBuffer } from 'lodash'
-/**
- * Check if valid Core URI
- *
- * @param {string} coreURI - The Core URI to test for validity
- * @returns {bool} true if coreURI is a valid Core URI, otherwise false
- */
-export function isValidCoreURI(coreURI) {
-  if (isEmpty(coreURI) || !isString(coreURI)) return false
-
-  try {
-    return isURL(coreURI, {
-      protocols: ['https'],
-      require_protocol: true,
-      host_whitelist: [/^[a-z]\.chainpoint\.org$/]
-    })
-  } catch (error) {
-    return false
-  }
-}
-
-/**
- * Check if valid Node URI
- *
- * @param {string} nodeURI - The value to check
- * @returns {bool} true if value is a valid Node URI, otherwise false
- */
-export function isValidNodeURI(nodeURI) {
-  if (!isString(nodeURI)) return false
-
-  try {
-    let isValidURI = isURL(nodeURI, {
-      protocols: ['http', 'https'],
-      require_protocol: true,
-      host_blacklist: ['0.0.0.0']
-    })
-
-    let parsedURI = url.parse(nodeURI).hostname
-
-    // Valid URI w/ IPv4 address?
-    return isValidURI && isIP(parsedURI, 4)
-  } catch (error) {
-    return false
-  }
-}
-
-/**
- * Checks if value is a hexadecimal string
- *
- * @param {string} value - The value to check
- * @returns {bool} true if value is a hexadecimal string, otherwise false
- */
-export function isHex(value) {
-  var hexRegex = /^[0-9a-f]{2,}$/i
-  var isHex = hexRegex.test(value) && !(value.length % 2)
-  return isHex
-}
-
-/**
- * Checks if a proof handle Object has valid params.
- *
- * @param {Object} handle - The proof handle to check
- * @returns {bool} true if handle is valid Object with expected params, otherwise false
- */
-export function isValidProofHandle(handle) {
-  if (!isEmpty(handle) && isObject(handle) && has(handle, 'uri') && has(handle, 'hashIdNode')) {
-    return true
-  }
-}
-
-/**
- * Checks if a UUID is a valid v1 UUID.
- *
- * @param {string} uuid - The uuid to check
- * @returns {bool} true if uuid is valid, otherwise false
- */
-export function isValidUUID(uuid) {
-  if (uuidValidate(uuid, 1)) {
-    return true
-  }
-}
-
-/**
- * Check if client is being used over an https connection
- * @returns {bool} true if served over https
- */
-export function isSecureOrigin() {
-  return typeof window === 'object' && window.location.protocol === 'https:'
-}
+import { isHex } from './helpers'
 
 /**
  * Map the JSON API response from submitting a hash to a Node to a
